Reduce public API surface
