+++ sql/mysql_test.go
@@ -13,7 +13,7 @@ func TestMySqlRepository_ListDatabases(t *testing.T) {
 	ctx, db, mock, r := initMySqlRepoTest(t)
 	defer func() { _ = db.Close() }()
 	dbRows := sqlmock.NewRows([]string{"name"}).AddRow("db1").AddRow("db2")
-	mock.ExpectQuery(MySqlDatabaseQuery).WillReturnRows(dbRows)
+	mock.ExpectQuery(mySqlDatabaseQuery).WillReturnRows(dbRows)
 	dbs, err := r.ListDatabases(ctx)
 	require.NoError(t, err)
 	require.ElementsMatch(t, []string{"db1", "db2"}, dbs)
